In proc changes

+++ DebtCollectionAccess/DebtCollectionAccess/Models/AccountBalance.cs
@@ -7,11 +7,11 @@ namespace DebtCollectionAccess
     public partial class AccountBalance
     {
         public int Id { get; set; }
-        public double OpeningBalance { get; set; }
-        public double? Paid { get; set; }
+        public decimal OpeningBalance { get; set; }
+        public decimal? Paid { get; set; }
         public int? PeriodId { get; set; }
-        public double? PromisedAmount { get; set; }
-        public double? RemainingBalance { get; set; }
+        public decimal? PromisedAmount { get; set; }
+        public decimal? RemainingBalance { get; set; }
         public int AccountId { get; set; }
         public bool? IsPaymentMissed { get; set; }
         public bool? IsPartialPayment { get; set; }
