+++ DebtCollectionAccess/DebtCollectionAccess/Models/Commission.cs
@@ -6,8 +6,8 @@ namespace DebtCollectionAccess
     public partial class Commission
     {
         public int Id { get; set; }
-        public double LowerRange { get; set; }
-        public double HigherRange { get; set; }
-        public double CommissionPercentage { get; set; }
+        public decimal LowerRange { get; set; }
+        public decimal HigherRange { get; set; }
+        public decimal CommissionPercentage { get; set; }
     }
 }
