+++ DebtCollectionAccess/DebtCollectionAccess/Models/CompanyDiscount.cs
@@ -6,7 +6,7 @@ namespace DebtCollectionAccess
     public partial class CompanyDiscount
     {
         public int Id { get; set; }
-        public double? Discount { get; set; }
+        public decimal? Discount { get; set; }
         public int? CompanyId { get; set; }
 
         public virtual Company Company { get; set; }
